+++ src/Components/Components/src/RenderTree/RenderBatch.cs
@@ -1,13 +1,14 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using Microsoft.AspNetCore.Components.RenderTree;
-
-namespace Microsoft.AspNetCore.Components.Rendering
+namespace Microsoft.AspNetCore.Components.RenderTree
 {
     /// <summary>
-    /// Describes a set of UI changes.
+    /// Types in the Microsoft.AspNetCore.Components.RenderTree are not recommended for use outside
+    /// of the Blazor framework. These types will change in a future release.
     /// </summary>
+    //
+    // Describes a set of UI changes.
     public readonly struct RenderBatch
     {
         /// <summary>
