+++ src/Components/Shared/src/ArrayBuilder.cs
@@ -7,7 +7,7 @@ using System.Diagnostics;
 using System.Runtime.CompilerServices;
 
 #if COMPONENTS_SERVER
-namespace Microsoft.AspNetCore.Components.Web.Rendering
+namespace Microsoft.AspNetCore.Components.Server.Circuits
 #else
 namespace Microsoft.AspNetCore.Components.RenderTree
 #endif
