+++ src/Components/Web/src/WebEventDescriptor.cs
@@ -1,12 +1,11 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using Microsoft.AspNetCore.Components.Rendering;
-
-namespace Microsoft.AspNetCore.Components.Web
+namespace Microsoft.AspNetCore.Components.RenderTree
 {
     /// <summary>
-    /// For framework use only.
+    /// Types in the Microsoft.AspNetCore.Components.RenderTree are not recommended for use outside
+    /// of the Blazor framework. These types will change in a future release.
     /// </summary>
     public sealed class WebEventDescriptor
     {
