+++ src/Components/test/E2ETest/ServerExecutionTests/ComponentHubReliabilityTest.cs
@@ -18,6 +18,7 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Testing;
 using Xunit;
 using Xunit.Abstractions;
+using Microsoft.AspNetCore.Components.RenderTree;
 
 namespace Microsoft.AspNetCore.Components.E2ETest.ServerExecutionTests
 {
